refactor account profiles

+++ rental/models.py
@@ -1,4 +1,5 @@
 import datetime
+from decimal import Decimal
 from django.db import models
 from django.contrib.auth.models import User
 from django.urls import reverse
@@ -29,7 +30,7 @@ class Car(models.Model):
     name = models.CharField(max_length=100)
     image = models.ImageField(upload_to='car_images/%Y/%m/%d/', blank=True)
     description = models.TextField()
-    daily_rent = models.DecimalField(max_digits=5, decimal_places=2)
+    daily_rent = models.FloatField()
     is_available = models.BooleanField(default=True)
     created = models.DateTimeField(auto_now_add=True)
     localization = models.CharField(max_length=100, blank=True, null=True)
